test: split EthereumConnection provider tests
